requirements.txt
tensorflow==2.9.1
pyyaml==6.0
tqdm==4.64.0
tk==0.1.0
pillow=9.1.1
